+++ mopub-sdk/mopub-sdk-base/src/main/java/com/mopub/common/privacy/PersonalInfoManager.java
@@ -238,6 +238,7 @@ public class PersonalInfoManager {
      *
      * @return true for in GDPR region, false for not in GDPR region, null for unknown
      */
+    @Nullable
     public Boolean gdprApplies() {
         if (mPersonalInfoData.isForceGdprApplies()) {
             return true;