+++ mopub-sdk/mopub-sdk-rewardedvideo/src/main/java/com/mopub/mobileads/MoPubRewardedVideoManager.java
@@ -762,7 +762,6 @@ public class MoPubRewardedVideoManager {
                     onRewardedVideoPlaybackErrorAction(currentlyShowingAdUnitId, errorCode);
                 }
             });
-            sInstance.rewardedAdsLoaders.markFail(currentlyShowingAdUnitId);
         }
         sInstance.mRewardedAdData.setCurrentlyShowingAdUnitId(null);
     }
@@ -770,6 +769,7 @@ public class MoPubRewardedVideoManager {
     private static void onRewardedVideoPlaybackErrorAction(@NonNull final String adUnitId, @NonNull final MoPubErrorCode errorCode) {
         Preconditions.checkNotNull(adUnitId);
         Preconditions.checkNotNull(errorCode);
+        sInstance.rewardedAdsLoaders.markFail(adUnitId);
         if (sInstance.mVideoListener != null) {
             sInstance.mVideoListener.onRewardedVideoPlaybackError(adUnitId, errorCode);
         }
@@ -830,11 +830,11 @@ public class MoPubRewardedVideoManager {
 
     private static void onRewardedVideoClosedAction(@NonNull final String adUnitId) {
         Preconditions.checkNotNull(adUnitId);
+        // remove adloader from map
+        sInstance.rewardedAdsLoaders.markPlayed(adUnitId);
         if (sInstance.mVideoListener != null) {
             sInstance.mVideoListener.onRewardedVideoClosed(adUnitId);
         }
-        // remove adloader from map
-        sInstance.rewardedAdsLoaders.markPlayed(adUnitId);
     }
 
     public static <T extends CustomEventRewardedAd>
@@ -940,9 +940,8 @@ public class MoPubRewardedVideoManager {
     }
 
     private static void logErrorNotInitialized() {
-        MoPubLog.e("MoPub rewarded ad was not initialized. You must call " +
-                "MoPub.initializeRewardedVideo() before loading or attempting " +
-                "to play rewarded ads.");
+        MoPubLog.e("MoPub rewarded ads must be initialized with an Activity Context " +
+                "before calling rewarded ads methods.");
     }
 
     /**