+++ mopub-sdk/src/test/java/com/mopub/mobileads/AdViewControllerTest.java
@@ -95,6 +95,7 @@ public class AdViewControllerTest {
         Shadows.shadowOf(activity).grantPermissions(android.Manifest.permission.ACCESS_NETWORK_STATE);
 
         MoPub.initializeSdk(activity, new SdkConfiguration.Builder("adunit").build(), null);
+        Reflection.getPrivateField(MoPub.class, "sSdkInitialized").setBoolean(null, true);
 
         mockPersonalInfoManager = mock(PersonalInfoManager.class);
         when(mockPersonalInfoManager.getPersonalInfoConsentStatus()).thenReturn(ConsentStatus.UNKNOWN);
@@ -144,7 +145,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void setUserDataKeywords_shouldNotSetKeywordIfNoUserConsent() throws Exception {
+    public void setUserDataKeywords_shouldNotSetKeywordIfNoUserConsent() {
         when(mockPersonalInfoManager.canCollectPersonalInformation()).thenReturn(false);
 
         subject.setUserDataKeywords("user_data_keywords");
@@ -153,7 +154,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void setUserDataKeywords_shouldSetUserDataKeywordsIfUserConsent() throws Exception {
+    public void setUserDataKeywords_shouldSetUserDataKeywordsIfUserConsent() {
         when(mockPersonalInfoManager.canCollectPersonalInformation()).thenReturn(true);
 
         subject.setUserDataKeywords("user_data_keywords");
@@ -163,7 +164,7 @@ public class AdViewControllerTest {
 
 
     @Test
-    public void generateAdUrl_shouldNotSetUserDataKeywordsIfNoUserConsent() throws Exception {
+    public void generateAdUrl_shouldNotSetUserDataKeywordsIfNoUserConsent() {
         when(mockPersonalInfoManager.canCollectPersonalInformation()).thenReturn(false);
 
         subject.setAdUnitId("abc123");
@@ -178,7 +179,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void generateAdUrl_shouldSetUserDataKeywordsIfUserConsent() throws Exception {
+    public void generateAdUrl_shouldSetUserDataKeywordsIfUserConsent() {
         when(mockPersonalInfoManager.canCollectPersonalInformation()).thenReturn(true);
         when(mockPersonalInfoManager.getPersonalInfoConsentStatus()).thenReturn(
                 ConsentStatus.EXPLICIT_YES);
@@ -195,7 +196,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void adDidFail_shouldScheduleRefreshTimer_shouldCallMoPubViewAdFailed() throws Exception {
+    public void adDidFail_shouldScheduleRefreshTimer_shouldCallMoPubViewAdFailed() {
         ShadowLooper.pauseMainLooper();
         assertThat(Robolectric.getForegroundThreadScheduler().size()).isEqualTo(0);
 
@@ -206,7 +207,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void adDidFail_withNullMoPubView_shouldNotScheduleRefreshTimer_shouldNotCallMoPubViewAdFailed() throws Exception {
+    public void adDidFail_withNullMoPubView_shouldNotScheduleRefreshTimer_shouldNotCallMoPubViewAdFailed() {
         ShadowLooper.pauseMainLooper();
         assertThat(Robolectric.getForegroundThreadScheduler().size()).isEqualTo(0);
 
@@ -220,7 +221,7 @@ public class AdViewControllerTest {
 
 
     @Test
-    public void scheduleRefreshTimer_shouldNotScheduleIfRefreshTimeIsNull() throws Exception {
+    public void scheduleRefreshTimer_shouldNotScheduleIfRefreshTimeIsNull() {
         response = response.toBuilder().setRefreshTimeMilliseconds(null).build();
         subject.onAdLoadSuccess(response);
         ShadowLooper.pauseMainLooper();
@@ -244,7 +245,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void scheduleRefreshTimerIfEnabled_shouldCancelOldRefreshAndScheduleANewOne() throws Exception {
+    public void scheduleRefreshTimerIfEnabled_shouldCancelOldRefreshAndScheduleANewOne() {
         response = response.toBuilder().setRefreshTimeMilliseconds(30).build();
         subject.onAdLoadSuccess(response);
         ShadowLooper.pauseMainLooper();
@@ -260,7 +261,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void scheduleRefreshTimer_shouldNotScheduleRefreshIfAutoRefreshIsOff() throws Exception {
+    public void scheduleRefreshTimer_shouldNotScheduleRefreshIfAutoRefreshIsOff() {
         response = response.toBuilder().setRefreshTimeMilliseconds(30).build();
         subject.onAdLoadSuccess(response);
 
@@ -275,7 +276,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void scheduleRefreshTimer_whenAdViewControllerNotConfiguredByResponse_shouldHaveDefaultRefreshTime() throws Exception {
+    public void scheduleRefreshTimer_whenAdViewControllerNotConfiguredByResponse_shouldHaveDefaultRefreshTime() {
         ShadowLooper.pauseMainLooper();
         assertThat(Robolectric.getForegroundThreadScheduler().size()).isEqualTo(0);
 
@@ -290,7 +291,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void setShouldAllowAutoRefresh_shouldSetCurrentAutoRefreshStatus() throws Exception {
+    public void setShouldAllowAutoRefresh_shouldSetCurrentAutoRefreshStatus() {
         assertThat(subject.getCurrentAutoRefreshStatus()).isTrue();
 
         subject.setShouldAllowAutoRefresh(false);
@@ -301,7 +302,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void pauseRefresh_shouldDisableAutoRefresh() throws Exception {
+    public void pauseRefresh_shouldDisableAutoRefresh() {
         assertThat(subject.getCurrentAutoRefreshStatus()).isTrue();
 
         subject.pauseRefresh();
@@ -309,7 +310,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void resumeRefresh_afterPauseRefresh_shouldEnableRefresh() throws Exception {
+    public void resumeRefresh_afterPauseRefresh_shouldEnableRefresh() {
         subject.pauseRefresh();
 
         subject.resumeRefresh();
@@ -317,7 +318,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void pauseAndResumeRefresh_withShouldAllowAutoRefreshFalse_shouldAlwaysHaveRefreshFalse() throws Exception {
+    public void pauseAndResumeRefresh_withShouldAllowAutoRefreshFalse_shouldAlwaysHaveRefreshFalse() {
         subject.setShouldAllowAutoRefresh(false);
         assertThat(subject.getCurrentAutoRefreshStatus()).isFalse();
 
@@ -340,7 +341,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void enablingAutoRefresh_afterLoadAd_shouldScheduleNewRefreshTimer() throws Exception {
+    public void enablingAutoRefresh_afterLoadAd_shouldScheduleNewRefreshTimer() {
 
         final AdViewController adViewControllerSpy = spy(subject);
 
@@ -350,7 +351,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void enablingAutoRefresh_withoutCallingLoadAd_shouldNotScheduleNewRefreshTimer() throws Exception {
+    public void enablingAutoRefresh_withoutCallingLoadAd_shouldNotScheduleNewRefreshTimer() {
         final AdViewController adViewControllerSpy = spy(subject);
 
         adViewControllerSpy.setShouldAllowAutoRefresh(true);
@@ -358,7 +359,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void disablingAutoRefresh_shouldCancelRefreshTimers() throws Exception {
+    public void disablingAutoRefresh_shouldCancelRefreshTimers() {
         response = response.toBuilder().setRefreshTimeMilliseconds(30).build();
         subject.onAdLoadSuccess(response);
         ShadowLooper.pauseMainLooper();
@@ -372,7 +373,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void trackImpression_shouldAddToRequestQueue() throws Exception {
+    public void trackImpression_shouldAddToRequestQueue() {
         subject.onAdLoadSuccess(response);
         subject.trackImpression();
 
@@ -388,7 +389,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void registerClick_shouldHttpGetTheClickthroughUrl() throws Exception {
+    public void registerClick_shouldHttpGetTheClickthroughUrl() {
         subject.onAdLoadSuccess(response);
 
         subject.registerClick();
@@ -402,14 +403,14 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void fetchAd_withNullMoPubView_shouldNotMakeRequest() throws Exception {
+    public void fetchAd_withNullMoPubView_shouldNotMakeRequest() {
         subject.cleanup();
         subject.fetchAd("adUrl", null);
         verify(mockRequestQueue, never()).add(any(MultiAdRequest.class));
     }
 
     @Test
-    public void loadAd_shouldNotLoadWithoutConnectivity() throws Exception {
+    public void loadAd_shouldNotLoadWithoutConnectivity() {
         ConnectivityManager connectivityManager = (ConnectivityManager) RuntimeEnvironment.application.getSystemService(Context.CONNECTIVITY_SERVICE);
         Shadows.shadowOf(connectivityManager.getActiveNetworkInfo()).setConnectionStatus(false);
         subject.setAdUnitId("adunit");
@@ -419,22 +420,23 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void loadAd_shouldNotLoadUrlIfAdUnitIdIsNull() throws Exception {
+    public void loadAd_shouldNotLoadUrlIfAdUnitIdIsNull() {
         subject.loadAd();
 
         verifyZeroInteractions(mockRequestQueue);
     }
 
     @Test
-    public void loadNonJavascript_shouldFetchAd() throws Exception {
+    public void loadNonJavascript_shouldFetchAd() {
         String url = "https://www.guy.com";
+        reset(mockRequestQueue);
         subject.loadNonJavascript(url, null);
 
         verify(mockRequestQueue).add(argThat(isUrl(url)));
     }
 
     @Test
-    public void loadNonJavascript_whenAlreadyLoading_shouldNotFetchAd() throws Exception {
+    public void loadNonJavascript_whenAlreadyLoading_shouldNotFetchAd() {
         String url = "https://www.guy.com";
         subject.loadNonJavascript(url, null);
         reset(mockRequestQueue);
@@ -444,7 +446,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void loadNonJavascript_shouldAcceptNullParameter() throws Exception {
+    public void loadNonJavascript_shouldAcceptNullParameter() {
         subject.loadNonJavascript(null, null);
         // pass
     }
@@ -461,13 +463,13 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void loadFailUrl_shouldAcceptNullErrorCode() throws Exception {
+    public void loadFailUrl_shouldAcceptNullErrorCode() {
         subject.loadFailUrl(null);
         // pass
     }
 
     @Test
-    public void loadFailUrl_whenFailUrlIsNull_shouldCallAdDidFail() throws Exception {
+    public void loadFailUrl_whenFailUrlIsNull_shouldCallAdDidFail() {
         response.toBuilder().setFailoverUrl(null).build();
         subject.loadFailUrl(MoPubErrorCode.INTERNAL_ERROR);
 
@@ -476,7 +478,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void setAdContentView_whenCalledFromWrongUiThread_shouldStillSetContentView() throws Exception {
+    public void setAdContentView_whenCalledFromWrongUiThread_shouldStillSetContentView() {
         final View view = mock(View.class);
         AdViewController.setShouldHonorServerDimensions(view);
         subject.onAdLoadSuccess(response);
@@ -501,7 +503,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void setAdContentView_whenCalledAfterCleanUp_shouldNotRemoveViewsAndAddView() throws Exception {
+    public void setAdContentView_whenCalledAfterCleanUp_shouldNotRemoveViewsAndAddView() {
         final View view = mock(View.class);
         AdViewController.setShouldHonorServerDimensions(view);
         subject.onAdLoadSuccess(response);
@@ -521,7 +523,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void setAdContentView_whenHonorServerDimensionsAndHasDimensions_shouldSizeAndCenterView() throws Exception {
+    public void setAdContentView_whenHonorServerDimensionsAndHasDimensions_shouldSizeAndCenterView() {
         View view = mock(View.class);
         AdViewController.setShouldHonorServerDimensions(view);
         subject.onAdLoadSuccess(response);
@@ -539,7 +541,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void setAdContentView_whenHonorServerDimensionsAndDoesntHaveDimensions_shouldWrapAndCenterView() throws Exception {
+    public void setAdContentView_whenHonorServerDimensionsAndDoesntHaveDimensions_shouldWrapAndCenterView() {
         response = response.toBuilder().setDimensions(null, null).build();
         View view = mock(View.class);
         AdViewController.setShouldHonorServerDimensions(view);
@@ -558,7 +560,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void setAdContentView_whenNotServerDimensions_shouldWrapAndCenterView() throws Exception {
+    public void setAdContentView_whenNotServerDimensions_shouldWrapAndCenterView() {
         subject.onAdLoadSuccess(response);
         View view = mock(View.class);
 
@@ -661,7 +663,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void loadCustomEvent_shouldCallMoPubViewLoadCustomEvent() throws Exception {
+    public void loadCustomEvent_shouldCallMoPubViewLoadCustomEvent() {
         Map serverExtras = mock(Map.class);
         String customEventClassName = "customEventClassName";
         subject.loadCustomEvent(mockMoPubView, customEventClassName, serverExtras);
@@ -670,7 +672,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void loadCustomEvent_withNullMoPubView_shouldNotCallMoPubViewLoadCustomEvent() throws Exception {
+    public void loadCustomEvent_withNullMoPubView_shouldNotCallMoPubViewLoadCustomEvent() {
         Map serverExtras = mock(Map.class);
         String customEventClassName = "customEventClassName";
         subject.loadCustomEvent(null, customEventClassName, serverExtras);
@@ -679,7 +681,7 @@ public class AdViewControllerTest {
     }
 
     @Test
-    public void loadCustomEvent_withNullCustomEventClassName_shouldCallMoPubViewLoadCustomEvent() throws Exception {
+    public void loadCustomEvent_withNullCustomEventClassName_shouldCallMoPubViewLoadCustomEvent() {
         Map serverExtras = mock(Map.class);
         subject.loadCustomEvent(mockMoPubView, null, serverExtras);